+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Initial Telegram bot integration module structure
+
 ## [0.9.1] - 2026-01-17
 
 ### Added